fix: Reworked archi from client side -> Moved all the .cpp to src and .hpp to include.
