+++ client/include/components/animation.hpp
