+++ client/include/components/camera.hpp
