+++ client/include/components/client.hpp
@@ -1,0 +1,5 @@
+#include "render.hpp"
+#include "camera.hpp"
+#include "audio.hpp"
+#include "animation.hpp"
+#include "input.hpp"
