+++ client/src/render.cpp
@@ -1,4 +1,4 @@
-#include "client/components/render.hpp"
+#include "client/include/components/render.hpp"
 
 namespace rtype::client::components {
     Renderable::Renderable(const std::string& texturePath, float width, float height, RenderLayer layer)
